+++ pyodide-build/pyodide_build/out_of_tree/venv.py
@@ -86,6 +86,7 @@ def get_pip_monkeypatch(venv_bin: Path) -> str:
                 print([
                     os.name,
                     sys.platform,
+                    platform.system(),
                     sys.implementation._multiarch,
                     sysconfig.get_platform()
                 ])
@@ -129,7 +130,7 @@ def get_pip_monkeypatch(venv_bin: Path) -> str:
             yield from tags._generic_platforms()
 
         def platform_tags():
-            if platform.system() == "emscripten":
+            if platform.system() == "Emscripten":
                 yield from _emscripten_platforms()
                 return
             return orig_platform_tags()
@@ -137,12 +138,12 @@ def get_pip_monkeypatch(venv_bin: Path) -> str:
         tags.platform_tags = platform_tags
         """
         f"""
-        os_name, sys_platform, multiarch, host_platform = {platform_data}
+        os_name, sys_platform, platform_system, multiarch, host_platform = {platform_data}
         os.name = os_name
         orig_platform = sys.platform
         sys.platform = sys_platform
         sys.implementation._multiarch = multiarch
-        platform.system = lambda: sys_platform
+        platform.system = lambda: platform_system
         platform.machine = lambda: "wasm32"
         os.environ["_PYTHON_HOST_PLATFORM"] = host_platform
         os.environ["_PYTHON_SYSCONFIGDATA_NAME"] = f'_sysconfigdata_{{sys.abiflags}}_{{sys.platform}}_{{sys.implementation._multiarch}}'